Changed design

+++ project/barreviews/views.py
@@ -9,12 +9,12 @@ from django.core.context_processors import csrf
 # Login views
 
 def login(request):
-    if request.user.username == '':
-        c = {}
-        c.update(csrf(request))
-        return render_to_response('login.html',c)
-    else:
-        return HttpResponseRedirect("/users/")
+	if request.user.username == '':
+		c = {}
+		c.update(csrf(request))
+		return render_to_response('login.html',c)
+	else:
+		return HttpResponseRedirect("/users/")
 
 def auth_view(request):
 	username = request.POST.get('username', '')
@@ -202,10 +202,10 @@ def user_edit(request, pk):
 	return render(request, 'barreviews/user_edit.html', {'form': form})
 
 def user_delete(request, pk):
-    if int(request.user.id) == int(pk):
-        instance = User.objects.get(pk=pk)
-        instance.delete()
-    return redirect('barreviews:users')
+	if int(request.user.id) == int(pk):
+		instance = User.objects.get(pk=pk)
+		instance.delete()
+	return redirect('barreviews:users')
 	
 class ReviewView(generic.DetailView):
 	model = ReviewBar
@@ -246,8 +246,10 @@ def review_edit(request, pk):
 
 def review_delete(request, pk):
 	instance = ReviewBar.objects.get(pk=pk)
-	instance.delete()
-	return redirect('barreviews:reviews')
+	temp = instance
+	if request.user.pk == instance.user.pk:
+		instance.delete()
+	return redirect('barreviews:bar', pk=temp.bar.id)
 	
 class BreweriesView(generic.ListView):
 	template_name = 'barreviews/breweries.html'
@@ -317,7 +319,8 @@ def comment_edit(request, pk):
 def comment_delete(request, pk):
 	instance = Comment.objects.get(pk=pk)
 	temp = instance
-	instance.delete()
+	if request.user.pk == instance.user1.pk:
+		instance.delete()
 	return redirect('barreviews:user', pk=temp.user2.pk)
 
 def comment_auth(request, pk):
